requirements.txt
azure-functions
fastapi
line-bot-sdk